% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{components}
\alias{components}
\alias{components,ICAResult-method}
\title{Extract Independent Components}
\usage{
components(object)

\S4method{components}{ICAResult}(object)
}
\arguments{
\item{object}{An ICAResult object}
}
\value{
Matrix of independent components (n.comp x n observations)
}
\description{
Extract Independent Components
}
