% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fastICA_dense}
\alias{cpp_fastICA_dense}
\title{Fast Independent Component Analysis (C++ Implementation)}
\usage{
cpp_fastICA_dense(
  X,
  n_comp,
  alg_type,
  fun_type,
  alpha,
  whiten_method,
  max_iter,
  tol,
  verbose,
  n_threads,
  seed
)
}
\arguments{
\item{X}{Numeric matrix (observations x variables)}

\item{n_comp}{Number of independent components to extract}

\item{alg_type}{Algorithm type: 0 = parallel (symmetric), 1 = deflation}

\item{fun_type}{Nonlinearity: 0 = logcosh, 1 = exp, 2 = cube}

\item{alpha}{Parameter for logcosh (between 1 and 2)}

\item{whiten_method}{Whitening: 0 = SVD, 1 = eigendecomposition}

\item{max_iter}{Maximum iterations}

\item{tol}{Convergence tolerance}

\item{verbose}{Print progress}

\item{n_threads}{Number of OpenMP threads (0 = auto)}

\item{seed}{Random seed (0 = random)}
}
\value{
List with S, A, W, K, center, iterations, converged
}
\description{
Fast Independent Component Analysis (C++ Implementation)
}
\keyword{internal}
