% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_whiten}
\alias{cpp_whiten}
\title{Whitening only (for testing/debugging)}
\usage{
cpp_whiten(X, n_comp, method)
}
\arguments{
\item{X}{Numeric matrix}

\item{n_comp}{Number of components}

\item{method}{0 = SVD, 1 = eigen}
}
\value{
List with X_whitened, K, K_inv, mean
}
\description{
Whitening only (for testing/debugging)
}
\keyword{internal}
