% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ICAResult-class}
\alias{ICAResult-class}
\title{ICAResult Class}
\description{
S4 class to store results from fastICA
}
\section{Slots}{

\describe{
\item{\code{S}}{Matrix of independent components (n.comp x n observations)}

\item{\code{A}}{Mixing matrix (variables x n.comp)}

\item{\code{W}}{Unmixing matrix (n.comp x whitened dimension)}

\item{\code{K}}{Whitening matrix (variables x whitened dimension)}

\item{\code{center}}{Numeric vector of column means used for centering}

\item{\code{misc}}{List containing metadata (iterations, converged, call, runtime, etc.)}
}}

