% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mixing}
\alias{mixing}
\alias{mixing,ICAResult-method}
\title{Extract Mixing Matrix}
\usage{
mixing(object)

\S4method{mixing}{ICAResult}(object)
}
\arguments{
\item{object}{An ICAResult object}
}
\value{
Mixing matrix (variables x n.comp)
}
\description{
Extract Mixing Matrix
}
