% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict,ICAResult-method}
\alias{predict,ICAResult-method}
\title{Predict Method for ICAResult}
\usage{
\S4method{predict}{ICAResult}(object, newdata)
}
\arguments{
\item{object}{An ICAResult object}

\item{newdata}{Optional new data matrix (observations x variables).
If missing, returns transposed S from original fit.}
}
\value{
Matrix of independent component scores (observations x n.comp)
}
\description{
Project new data onto learned independent components
}
\examples{
\dontrun{
# Fit ICA
result <- fastICA(X_train, n.comp = 5)

# Project new data
S_new <- predict(result, X_test)
}

}
