% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,ICAResult-method}
\alias{show,ICAResult-method}
\title{Show Method for ICAResult}
\usage{
\S4method{show}{ICAResult}(object)
}
\arguments{
\item{object}{An ICAResult object}
}
\description{
Show Method for ICAResult
}
