% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,ICAResult-method}
\alias{summary,ICAResult-method}
\title{Summary Method for ICAResult}
\usage{
\S4method{summary}{ICAResult}(object)
}
\arguments{
\item{object}{An ICAResult object}
}
\value{
Invisibly returns the object
}
\description{
Summary Method for ICAResult
}
